CO: Add PHP 7.1 support

+++ controllers/admin/AdminImportController.php
@@ -648,7 +648,7 @@ class AdminImportControllerCore extends AdminController
 
         //get post max size
         $post_max_size = ini_get('post_max_size');
-        $bytes         = trim($post_max_size);
+        $bytes         = (int) trim($post_max_size);
         $last          = strtolower($post_max_size[strlen($post_max_size) - 1]);
 
         switch ($last) {
